+++ CP2exec/CP2ExecThread.h
@@ -21,10 +21,10 @@ public:
 	void pulses(int& pulses1, int& pulses2, int& pulses);
 	void rates(double& rate1, double& rate2, double& rate3);
 	void eof(bool eof[3]);
-	void antennaInfo(unsigned short* az, 
-		unsigned short* el, 
-		unsigned short* sweep, 
-		unsigned short* volume);
+	void antennaInfo(double* az, 
+		double* el, 
+		int* sweep, 
+		int* volume);
 
 protected:
 	/// The piraq dsp's will read the antenna pointing information
